small bugs fixes & changes. moved jwt token related stuff to utils/auth
